backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
psycopg[binary]==3.1.18
Flask-Babel==4.0.0
python-dotenv==1.0.1
PyMySQL==1.1.1
